pyproject.toml
[tool.poetry]
name = "habbits-api"
version = "0.1.0"
description = ""
authors = ["Goldyban Nastya <nasty-goldyba@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.4.0"
drf-spectacular = "^0.28.0"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.1"
eventlet = "^0.39.0"
django-cors-headers = "^4.7.0"
requests = "^2.32.3"


[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.6.12"

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
